Move AppContext to package "app"
