+++ src/main/java/org/example/controller/LoginController.java
@@ -8,7 +8,7 @@ import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 import javafx.scene.control.PasswordField;
 import javafx.scene.control.TextField;
-import org.example.model.AppContext;
+import org.example.app.AppContext;
 import org.example.model.dto.AuthRequest;
 import org.example.service.AuthServiceImpl;
 import org.example.app.Main;
